requirements.txt
torch
torchvision
torchaudio
streamlit
pillow
pytest
